server: erlang: add payload support on the dispatcher
The WRT command uses a payload defined by a SIZE N directive. We should
not be looking for a newline inside of it.

For example, the first newline on "WRT FD 1 SIZE 11 HELLO\nWORLD\n" does
not indicate a line break on the command stream. The second one, on the
other hand, does indicate so.